[ANDROSDK-460] Move PerformanceHintsServiceShould to maintenance
